skypulse-email/requirements.txt
# SkyPulse 2.0 - Email Service Dependencies

# Core
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Email
email-validator==2.1.0

# AI/LLM
ollama==0.1.6
groq==0.4.2

# Utilities
requests==2.31.0
python-dateutil==2.8.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Scheduling
apscheduler==3.10.4

# Optional: Web UI
flask==3.0.0
flask-cors==4.0.0

# Development
pytest==7.4.3
black==23.12.1
